查询点赞列表初步完成

+++ model/video.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Video struct {
 	gorm.Model
-	Uid           int64  `json:"uid"`                                // 用户id
+	Uid           uint   `json:"uid"`                                // 用户id
 	PlayUrl       string `json:"play_url" json:"play_url,omitempty"` // 视频播放地址
 	CoverUrl      string `json:"cover_url,omitempty"`                // 视频封面地址
 	Title         string `json:"title,omitempty"`                    // 视频标题
